+++ lib/src/plugin/mod.rs
@@ -1,5 +1,7 @@
 mod convert_py_data;
 mod definition;
+#[cfg(feature = "plugin")]
+pub mod exec_processor;
 mod init_python;
 #[allow(clippy::module_name_repetitions)]
 pub mod plugin_manager;