+++ lib/src/response/handler/arrow_parquet.rs
@@ -21,7 +21,7 @@ pub async fn parquet_responder(
     let record_batches = parquet::to_record_batch(file_path.to_str().unwrap())?;
 
     Ok(http_response::from_byte_stream(
-        arrow_stream::make_stream(&record_batches)
+        arrow_stream::make_buffered_stream(&record_batches)
             .map_err(ResponseError::arrow_stream_serialization)?,
         "application/vnd.apache.arrow.stream",
     ))