+++ lib/src/response/mod.rs
@@ -3,3 +3,5 @@ pub mod handler;
 pub mod http_error;
 #[allow(clippy::module_name_repetitions)]
 pub mod http_response;
+pub mod receiver_stream;
+pub mod record_batch_stream;